+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * :fire: Filex mkdirsCheck rename to mkdirsOrCheck, createNewFileCheck rename to createNewFileOrCheck
 * :sparkles: Filex adds clean method
 * :fire: Filex listFilesRecursively(File, @Nullable FileFilter) method change to listFilesRecursively(File, @NotNull FileFilter), listFilesRecursively(File, @Nullable FilenameFilter) method change to listFilesRecursively(File, @NotNull FilenameFilter)
+* :sparkles: Filex adds `String[] listRecursively(File dir, FilenameFilter)` and `String[] listRecursively(File dir, FileFilter)` method
 
 
 ## v0.8.3